advanced-bedrock-deployment/fraud-detection-agent/pyproject.toml
[project]
name = "fraud-detection-agent-advanced"
version = "2.0.0"
description = "Advanced Strands SDK Fraud Detection Agent with AgentCore Memory"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "bedrock-agentcore>=0.0.1",
    "boto3>=1.34.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.0.0",
]
